Move help to General and make commands bold
